Migrate data transformation into `InteractsWithData` trait
